RR-Team_66-yadtq-yet-another-distributed-task-queue-/final/requirements.txt
#requirements.txt
kafka-python
redis
uuid
python-dotenv
